requirements.txt
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.28.2
six==1.16.0
stripe==5.4.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
